+++ react/__tests__/BankInvoiceSection.test.tsx
@@ -4,6 +4,7 @@ import { orderGroupQuery as bankInvoiceLoggedIn } from '../mocks/bankInvoiceLogg
 import { orderGroupQuery as bankInvoiceNotLoggedIn } from '../mocks/bankInvoiceNotLoggedIn'
 import { orderGroupQuery as bankInvoiceNumberLoggedIn } from '../mocks/bankInvoiceNumberLoggedIn'
 import { orderGroupQuery as bankInvoiceNumberNotLoggedIn } from '../mocks/bankInvoiceNumberNotLoggedIn'
+import { orderGroupQuery as bankInvoiceDisplayNonBankUrlPayment } from '../mocks/bankInvoiceDisplayNonBankUrlPayment'
 import BankInvoiceSection from '../BankInvoiceSection'
 import { renderWithOrderGroup } from '../utils/testUtils'
 
@@ -75,4 +76,15 @@ test('has the invoice url if the user is logged in.', () => {
 
   const printButton = container.querySelector('a[href]') as HTMLAnchorElement
   expect(printButton.href).not.toContain('login')
+})
+
+test('renders embedded for other types of payments besides bank invoice', () => {
+  const { queryByTestId } = renderWithOrderGroup(
+    bankInvoiceDisplayNonBankUrlPayment.orderGroup,
+    <BankInvoiceSection />
+  )
+
+  const embedded = queryByTestId('embedded-bank-invoice')
+
+  expect(embedded).toBeTruthy()
 })