+++ CHANGELOG.md
@@ -7,18 +7,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [2.4.2] - 2020-08-04
+### Fixed
+
+- Display of embedded info for non `bankIvoice` payments.
 
 ## [2.4.2] - 2020-08-04
+
 ### Added
+
 - Fixed display of images on the website.
 
 ## [2.4.1] - 2020-08-03
+
 ### Added
+
 - IO app typings.
 
 ## [2.4.0] - 2020-08-03
+
 ### Added
+
 - New css handles: `barCodeContainer`, `printButtonWrapper` and `printHintWrapper`.
 
 ## [2.3.0] - 2020-07-08
@@ -30,7 +38,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.2.5] - 2020-06-22
 
 ## [2.2.4] - 2020-06-22
+
 ### Fixed
+
 - `main` tag not working properly in IE 11.
 - Possible duped keys in `OrderTotal`.